+++ lib/zaikio/client/helpers/locale_middleware.rb
@@ -11,20 +11,10 @@ module Zaikio
 
         def call(request_env)
           if !@disable_i18n && Object.const_defined?("I18n")
-            request_env.url.query = add_query_param(request_env.url.query, "lang", ::I18n.try(:locale))
+            request_env[:request_headers]["Accept-Language"] = ::I18n.try(:locale)&.to_s
           end
 
           @app.call(request_env)
-        end
-
-        private
-
-        def add_query_param(query, key, value)
-          query = [query.to_s]
-          query << "&" unless query.empty?
-          query << "#{Faraday::Utils.escape key}=#{Faraday::Utils.escape value}"
-
-          query.join
         end
       end
     end