+++ test/zaikio/client/helpers/locale_middleware_test.rb
@@ -19,32 +19,32 @@ class Zaikio::Client::Helpers::LocaleMiddlewareTest < ActiveSupport::TestCase
   test "it sets the lang query param" do
     @connection = create_connection
     
-    stub_request(:get, "https://procurement.sandbox.zaikio.com/api/v2/resources?anotherparam=1&lang=de")
+    stub_request(:get, "https://procurement.sandbox.zaikio.com/api/v2/resources")
 
     I18n.with_locale(:de) do
       Zaikio::Client.with_token("token") do
-        @connection.get("resources?anotherparam=1")
+        @connection.get("resources")
       end
     end
 
     assert_requested(
-      :get, "https://procurement.sandbox.zaikio.com/api/v2/resources?anotherparam=1&lang=de",
-      headers: { "Authorization" => "Bearer token" }
+      :get, "https://procurement.sandbox.zaikio.com/api/v2/resources",
+      headers: { "Authorization" => "Bearer token", "Accept-Language" => "de" }
     )
   end
 
   test "it does not set the lang query param if disabled" do
     @connection = create_connection(disable_i18n: true)
-    stub_request(:get, "https://procurement.sandbox.zaikio.com/api/v2/resources?anotherparam=1")
+    stub_request(:get, "https://procurement.sandbox.zaikio.com/api/v2/resources")
 
     I18n.with_locale(:de) do
       Zaikio::Client.with_token("token") do
-        @connection.get("resources?anotherparam=1")
+        @connection.get("resources")
       end
     end
 
     assert_requested(
-      :get, "https://procurement.sandbox.zaikio.com/api/v2/resources?anotherparam=1",
+      :get, "https://procurement.sandbox.zaikio.com/api/v2/resources",
       headers: { "Authorization" => "Bearer token" }
     )
   end